added routes and handlebars
