+++ controllers/api/eventRoutes.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { Event } = require('../../models');
+const { Event } = require('../../loginsw/models');
 const withAuth = require('../../utils/auth');
 
 router.post('/', withAuth, async (req, res) => {
@@ -25,7 +25,7 @@ router.delete('/:id', withAuth, async (req, res) => {
     });
 
     if (!eventData) {
-      res.status(404).json({ message: 'No event found with this id!' });
+      res.status(404).json({ message: 'No event found!' });
       return;
     }
