+++ controllers/index.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 
-const apiRoutes = require('./api');
+const apiRoutes = require('../controllers');
 const homeRoutes = require('./homeRoutes');
 
 router.use('/', homeRoutes);
