--- loginsw/db/schema.sql
@@ -1,2 +1,0 @@
-DROP DATABASE IF EXISTS crowdfund_db;
-CREATE DATABASE crowdfund_db;
