+++ loginsw/models/index.js
@@ -1,13 +1,15 @@
-const User = require('./User');
-const Event = require('./Event');
+// const User = require('./User');
+// const Event = require('./Event');
 
-User.hasMany(Event, {
-  foreignKey: 'user_id',
-  onDelete: 'CASCADE'
-});
+// User.hasMany(Event, {
+//   foreignKey: 'user_id',
+//   onDelete: 'CASCADE'
+// });
 
-Event.belongsTo(User, {
-  foreignKey: 'user_id'
-});
+// Event.belongsTo(User, {
+//   foreignKey: 'user_id'
+// });
 
-module.exports = { User, Event };
+// module.exports = { User, Event };
+
+// user_id changed to creator_id
