+++ loginsw/public/js/login.js
@@ -1,63 +1,63 @@
-const loginFormHandler = async (event) => {
-  event.preventDefault();
+// const loginFormHandler = async (event) => {
+//   event.preventDefault();
 
-  // Collect values from the login form
-  const email = document.querySelector('#email-login').value.trim();
-  const password = document.querySelector('#password-login').value.trim();
+//   // Collect values from the login form
+//   const email = document.querySelector('#email-login').value.trim();
+//   const password = document.querySelector('#password-login').value.trim();
 
-  if (email && password) {
-    // Send a POST request to the API endpoint
-    const response = await fetch('/api/users/login', {
-      method: 'POST',
-      body: JSON.stringify({ email, password }),
-      headers: { 'Content-Type': 'application/json' },
-    });
+//   if (email && password) {
+//     // Send a POST request to the API endpoint
+//     const response = await fetch('/api/users/login', {
+//       method: 'POST',
+//       body: JSON.stringify({ email, password }),
+//       headers: { 'Content-Type': 'application/json' },
+//     });
 
-    if (response.ok) {
-      // If successful, redirect the browser to the profile page
-      document.location.replace('/profile');
-    } else {
-      alert(response.statusText);
-    }
-  }
-};
+//     if (response.ok) {
+//       // If successful, redirect the browser to the profile page
+//       document.location.replace('/profile');
+//     } else {
+//       alert(response.statusText);
+//     }
+//   }
+// };
 
-const signupFormHandler = async (event) => {
-  event.preventDefault();
+// const signupFormHandler = async (event) => {
+//   event.preventDefault();
 
-  const name = document.querySelector('#name-signup').value.trim();
-  const email = document.querySelector('#email-signup').value.trim();
-  const password = document.querySelector('#password-signup').value.trim();
-  const isOrganizer = document.querySelector('#organizer-signup').checked;
+//   const name = document.querySelector('#name-signup').value.trim();
+//   const email = document.querySelector('#email-signup').value.trim();
+//   const password = document.querySelector('#password-signup').value.trim();
+//   const isOrganizer = document.querySelector('#organizer-signup').checked;
 
-  if (name && email && password) {
-    const userData = {
-      name,
-      email,
-      password,
-      isOrganizer,
-    };
+//   if (name && email && password) {
+//     const userData = {
+//       name,
+//       email,
+//       password,
+//       isOrganizer,
+//     };
 
-    const response = await fetch('/api/users', {
-      method: 'POST',
-      body: JSON.stringify(userData),
-      headers: { 'Content-Type': 'application/json' },
-    });
+//     const response = await fetch('/api/users', {
+//       method: 'POST',
+//       body: JSON.stringify(userData),
+//       headers: { 'Content-Type': 'application/json' },
+//     });
 
-    if (response.ok) {
-      document.location.replace('/profile');
-    } else {
-      alert(response.statusText);
-    }
-  }
-};
+//     if (response.ok) {
+//       document.location.replace('/profile');
+//     } else {
+//       alert(response.statusText);
+//     }
+//   }
+// };
 
 
 
-document
-  .querySelector('.login-form')
-  .addEventListener('submit', loginFormHandler);
+// document
+//   .querySelector('.login-form')
+//   .addEventListener('submit', loginFormHandler);
 
-document
-  .querySelector('.signup-form')
-  .addEventListener('submit', signupFormHandler);
+// document
+//   .querySelector('.signup-form')
+//   .addEventListener('submit', signupFormHandler);
