--- loginsw/public/js/logout.js
@@ -1,14 +1,0 @@
-const logout = async () => {
-  const response = await fetch('/api/users/logout', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-  });
-
-  if (response.ok) {
-    document.location.replace('/');
-  } else {
-    alert(response.statusText);
-  }
-};
-
-document.querySelector('#logout').addEventListener('click', logout);
