+++ views/event.handlebars
@@ -4,8 +4,8 @@
   <p>{{description}}</p>
   <p>Created by {{user.name}} on {{format_date date_created}}</p>
   <p>
-    <span for="img" aria-label="money">💲</span>
-    <span class="dollar-amount">{{format_amount needed_funding}}</span>
+    <span for="img" aria-label="person">💲</span>
+    <span class="volunteer-amount">{{format_amount needed_volunteers}}</span>
     needed
   </p>
   <p>
