+++ views/profile.handlebars
@@ -7,6 +7,9 @@
 {{#if isOrganizer}}
 <div class="row mt-4">
   <div class="col-md-6">
+
+
+    
     <h3>Create a New Event:</h3>
 
     <form class="form new-event-form">
@@ -15,8 +18,8 @@
         <input class="form-input" type="text" id="event-name" name="event-name" />
       </div>
       <div class="form-group">
-        <label for="event-funding">Number of Volunteers Needed:</label>
-        <input class="form-input" type="number" id="event-funding" name="event-funding" />
+        <label for="event-volunteers">Number of Volunteers Needed:</label>
+        <input class="form-input" type="number" id="event-volunteers" name="event-volunteers" />
       </div>
       <div class="form-group">
         <label for="event-desc">Event description:</label>
